coordinates converted from float to double

+++ RODAS/src/core/Point.java
@@ -2,7 +2,7 @@ package core;
 
 public class Point {
 
-	public Point(int id, float x, float y, float z) {
+	public Point(int id, double x, double y, double z) {
 		super();
 		this.id = id;
 		this.x  = x;
@@ -11,9 +11,9 @@ public class Point {
 	}
 
 	int   id;
-	float x;
-	float y;
-	float z;
+	double x;
+	double y;
+	double z;
 	
 	public int getId() {
 		return id;
@@ -23,27 +23,27 @@ public class Point {
 		this.id = id;
 	}
 	
-	public float getX() {
+	public double getX() {
 		return x;
 	}
 	
-	public void setX(float x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 	
-	public float getY() {
+	public double getY() {
 		return y;
 	}
 	
-	public void setY(float y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 	
-	public float getZ() {
+	public double getZ() {
 		return z;
 	}
 	
-	public void setZ(float z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
